// src/enums/EnumsByFilter.ts
export enum EnumRuleAutofilter {
  equally = 'Равно',
  morethan = 'Больше чем',
  lessthan = 'Меньше чем',

  absent = 'Отсутствует',
  notequalto = 'Не равно',

  before = 'До',
  after = 'После',
  between = 'Между',

  tomorrow = 'Завтра',
  today = 'Сегодня',
  yesterday = 'Вчера',
  upthisweek = 'На этой неделе',
  thismonth = 'В этом месяце',
  thisyear = 'В этом году'
}
export enum EnumStepNumberInput {
  One = 1,
  Ten = 10,
  Hundred = 100,
  Thousand = 1000
}  
export enum EnumTypeField {
  String = 'string',
  Number = 'number',
  Date = 'date',
  Other = 'other',
  Boolean = 'boolean',
  NoType = 'noType'
}  
export enum EnumTypeCell {
  String = 'String',
  Number = 'Number',
  Date = 'Date',
  Boolean = 'Boolean',
  Dropdown = 'Dropdown'
}  
export enum EnumTypeFilter {
  FastFilter = 'FastFilter',
  CommonFilter = 'CommonFilter'
} 